Prevent encoding the index name
In Elasticsearch, the first parameter of the path is the index name. When using single-index search without special characters everything works fine.

As soon as you try to use multi-index search or your index name contains special characters everything breaks.

This happens because in lines 304-307 we encode the full url. The same code is used in other AWS services so I am not sure if there is specific reason for doing so.

In ES though, the first parameter is the index name. When you use multi-index search the requested path should be something like this `categories,products/_msearch` and after encoding it becomes `categories%2Cproducts/_msearch` which fails.

Instead we just skip from encoding the first part for now.